requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# ── Redis ────────────────────────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Auth & Security ──────────────────────────────────────────────────────────
bcrypt==4.2.1
PyJWT==2.10.1

# ── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.28.1

# ── LLM ──────────────────────────────────────────────────────────────────────
openai==1.59.7

# ── Stripe ───────────────────────────────────────────────────────────────────
stripe==11.4.1

# ── SSE ──────────────────────────────────────────────────────────────────────
sse-starlette==2.2.1

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# ── Dev Tools ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

email-validator==2.2.0
# ── PDF Generation ───────────────────────────────────────────────────────────
fpdf2==2.8.2
